refactor: Refactor imports to aliases usage
