+++ packages/client/src/api/baseAxiosInstance.ts
@@ -2,7 +2,7 @@ import axios from 'axios';
 
 import { API_BASE_URL } from '@constants';
 
-export const instance = axios.create({
+export const baseAxiosInstance = axios.create({
   baseURL: API_BASE_URL,
   withCredentials: true,
 });
