--- packages/client/src/components/UnexpectedError/index.ts
@@ -1,1 +1,0 @@
-export { UnexpectedError } from './UnexpectedError';
