+++ packages/client/src/pages/Game.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react';
 
-import { Header } from '../components/Header';
+import { Header } from '@components/Header';
 
 export const Game: FC = () => {
   return (
