+++ packages/client/src/pages/GameStart/GameStart.tsx
@@ -1,10 +1,11 @@
 import { FC } from 'react';
 import { Link } from 'react-router-dom';
 
-import { GAME_NAME } from '@src/constants';
+import { Header } from '@components/Header';
 
-import { Header } from '../../components/Header';
-import { RoutePaths } from '../../providers/Router/AppRouter/constants';
+import { GAME_NAME } from '@constants';
+
+import { RoutePaths } from '@src/providers/Router/AppRouter/constants';
 
 export const GameStart: FC = () => {
   return (
