+++ packages/client/src/pages/Profile.tsx
@@ -1,16 +1,15 @@
 import { FC } from 'react';
 
-import { authApi } from '../api/Auth/Auth';
-import { Header } from '../components/Header';
+import { Header } from '@components/Header';
 
-const logout = () => authApi.SignOut();
+import { authApi } from '@src/api/Auth/Auth';
 
 export const Profile: FC = () => {
   return (
     <>
       <Header />
       <h1 className="text-4xl font-bold">Profile</h1>
-      <button className="text-white btn-primary" onClick={logout}>
+      <button className="text-white btn-primary" onClick={authApi.SignOut}>
         Выйти из профиля
       </button>
     </>
